+++ TheCircleBackend/TheCircleBackend/Domain/DTO/EncryptedPayload/AuthUserContentDTO.cs
@@ -1,8 +1,8 @@
 ﻿using TheCircleBackend.Domain.Interfaces;
 
-namespace TheCircleBackend.Domain.DTO
+namespace TheCircleBackend.Domain.DTO.EncryptedPayload
 {
-    public class AuthUserContentDTO: IContent
+    public class AuthUserContentDTO : IContent
     {
         public LoginDTO OriginalLoginData { get; set; }
     }
